requirements.txt
paho-mqtt==2.1.0
fastapi==0.115.0
uvicorn==0.30.6
